Refactor(DTO): 중간지점 구하는 API 에서 DTO 의 name 필드들 모두 stationName 으로 변경
