+++ src/main/java/podo/odeego/web/security/oauth2/OAuth2LoginResponse.java
@@ -1,6 +1,6 @@
 package podo.odeego.web.security.oauth2;
 
-import podo.odeego.domain.member.dto.MemberJoinResponse;
+import podo.odeego.domain.member.entity.dto.MemberJoinResponse;
 
 public record OAuth2LoginResponse(
 	String accessToken,
